BitPin/requirements.txt
# Core dependencies
Django==4.2.5
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter for Python

# Celery and task queue
celery==5.3.4
redis==5.0.0

# Additional utilities
django-cors-headers==4.0.0  # For handling CORS
django-filter==23.2  # For filtering querysets in DRF

# Testing
pytest==7.4.0
pytest-django==4.5.2

# Linting and code quality
flake8==6.1.0
black==23.9.1

# Optional: For deployment
gunicorn==21.2.0  # WSGI HTTP server

# For securing database
python-decouple==3.8

# Math data manipulating
numpy==2.1.3

# Networking and async tasks
requests==2.32.3
aiohttp==3.11.9
asyncio==3.4.3